srcs/backend/requirements.txt
Django==4.2
django-crispy-forms==1.12.0
django-bootstrap4==24.3
djangorestframework==3.15
channels==4.1.0
asgiref==3.8.1
django-extensions==3.2.1
psycopg2-binary==2.9.3
django-filter==22.1
djangorestframework-jsonapi==7.0.2
inflection==0.5.1
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
django-cors-headers==4.4.0